+++ scripts/compile-scss.js
@@ -41,7 +41,7 @@ async function compileScssFiles({
 
   const inputFilenames = (await glob(sourcePattern, undefined)).filter(filename => {
     if (targetTheme == null) return true;
-    return filename === `src/theme_${targetTheme}.scss`;
+    return filename === `src/themes/amsterdam/theme_${targetTheme}.scss`;
   });
 
   await Promise.all(
@@ -147,7 +147,14 @@ if (require.main === module) {
   }
 
   compileScssFiles({
-    sourcePattern: path.join('src', 'theme_*.scss'),
+    sourcePattern: path.join('src/themes/legacy', 'legacy_*.scss'), 
+    destinationDirectory: 'dist',
+    docsVariablesDirectory: 'src-docs/src/views/theme/_json',
+    packageName: euiPackageName
+  });
+
+  compileScssFiles({
+    sourcePattern: path.join('src/themes/amsterdam', 'theme_*.scss'), 
     destinationDirectory: 'dist',
     docsVariablesDirectory: 'src-docs/src/views/theme/_json',
     packageName: euiPackageName