+++ src/services/index.ts
@@ -125,6 +125,7 @@ export { EuiWindowEvent } from './window_event';
 
 export {
   useCombinedRefs,
+  useUpdateEffect,
   useDependentState,
   useIsWithinBreakpoints,
   useMouseMove,
@@ -133,28 +134,4 @@ export {
 
 export { throttle } from './throttle';
 
-export {
-  EuiSystemContext,
-  EuiThemeContext,
-  EuiModificationsContext,
-  EuiColorModeContext,
-  useEuiTheme,
-  withEuiTheme,
-  WithEuiThemeProps,
-  EuiThemeProvider,
-  buildTheme,
-  computed,
-  isInverseColorMode,
-  getColorMode,
-  getComputed,
-  getOn,
-  mergeDeep,
-  setOn,
-  Computed,
-  ComputedThemeShape,
-  EuiThemeColorMode,
-  EuiThemeComputed,
-  EuiThemeModifications,
-  EuiThemeShape,
-  EuiThemeSystem,
-} from './theme';
+export * from './theme';